+++ src/components/Card/index.tsx
@@ -1,5 +1,4 @@
 import { Button, Card as C } from "antd";
-import { PlusOutlined, PicLeftOutlined } from "@ant-design/icons";
 import loadable from "@loadable/component";
 import useModal from "@/hooks/useModal";
 const CardDetail = loadable(() => import("@components/Modals/CardDetail"));
@@ -25,12 +24,6 @@ const Card = ({ text }: Props) => {
       <C style={S.Card} onClick={onClickCard} bodyStyle={S.Body}>
         <p>{text}</p>
       </C>
-      <S.ButtonWrapper>
-        <Button type="ghost" block icon={<PlusOutlined />} style={{ textAlign: "left", padding: 0 }}>
-          Add a card
-        </Button>
-        <Button type="ghost" block icon={<PicLeftOutlined />} style={S.TemplateButton}></Button>
-      </S.ButtonWrapper>
     </S.Container>
   );
 };
