+++ src/components/CardList/index.tsx
@@ -1,5 +1,5 @@
-import { Card as C } from "antd";
-import { DashOutlined } from "@ant-design/icons";
+import { Button, Card as C } from "antd";
+import { EllipsisOutlined, PlusOutlined, PicLeftOutlined } from "@ant-design/icons";
 import Card from "@components/Card";
 import * as S from "./style";
 
@@ -10,11 +10,17 @@ const CardList = () => {
         title="Card list"
         bordered={false}
         style={{ width: 300 }}
-        extra={<DashOutlined />}
+        extra={<EllipsisOutlined />}
         headStyle={S.Header}
         bodyStyle={S.Body}
       >
         <Card text="card" />
+        <S.ButtonWrapper>
+          <Button type="ghost" block icon={<PlusOutlined />} style={{ textAlign: "left", padding: 0 }}>
+            Add a card
+          </Button>
+          <Button type="ghost" block icon={<PicLeftOutlined />} style={S.TemplateButton}></Button>
+        </S.ButtonWrapper>
       </C>
     </S.Container>
   );
